requirements.txt
fastapi>=0.109.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.32.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
GitPython>=3.1.40,<4.0.0
httpx>=0.26.0,<0.28.0

# Static analysis & tests (also used at runtime by services)
pylint>=3.0.0,<4.0.0
mypy>=1.8.0,<2.0.0
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.23.0,<0.25.0

# Dev / testing
pytest-cov>=4.1.0,<6.0.0
